ClippingBoxコンポーネントをファイルに切り出す
伴って、
ClippingBoxesをExhibitionにリネームする。
LayerまでをExhibitionに含める。
